+++ src/Facades/Payment.php
@@ -8,6 +8,7 @@ use Illuminate\Support\Facades\Facade;
 /**
  * @method static \Shah\Parakit\Contracts\PaymentGateway driver(?string $name = null)
  * @method static void extend(string $driver, \Closure $creator)
+ * @method static void resolveMerchantUsing(\Closure $resolver)
  * @method static \Shah\Parakit\PaymentBuilder for(object|string $reference, string $keyAttribute = 'id')
  */
 class Payment extends Facade